convert dao to kotlin

+++ app/src/main/java/com/sjn/stamp/db/dao/BaseDao.java
@@ -4,11 +4,11 @@ import io.realm.Realm;
 
 public abstract class BaseDao {
 
-    protected Integer getAutoIncrementId(Realm realm, Class clazz) {
+    protected Long getAutoIncrementId(Realm realm, Class clazz) {
         Number maxId = realm.where(clazz).max("id");
         if (maxId != null) {
-            return maxId.intValue() + 1;
+            return maxId.intValue() + 1L;
         }
-        return 1;
+        return 1L;
     }
 }
